[nn] replace energonai.nn with colossalai.nn (#147)
* replace energonai.nn with colossalai.nn

* fix lm head

+++ energonai/model/endecoder.py
@@ -2,7 +2,7 @@ from typing import Callable
 import torch
 from torch import dtype
 from torch import nn
-from energonai.nn import LayerNorm1D
+from colossalai.nn import LayerNorm1D
 
 from .mlp import MLP1D
 from .attention import MultiHeadAttention1D
@@ -21,11 +21,11 @@ class Block1D(nn.Module):
                  max_seq_len: int = 512,
                  fused_qkv: bool = True,
                  is_decoder: bool = True,
-                 disable_past_cache = False) -> None:
+                 disable_past_cache=False) -> None:
         super().__init__()
 
         self.apply_post_layernorm = apply_post_layernorm
-        self.norm1 = LayerNorm1D(hidden_size, eps=layernorm_epsilon)
+        self.norm1 = LayerNorm1D(hidden_size, eps=layernorm_epsilon, dtype=dtype)
 
         self.attn = MultiHeadAttention1D(hidden_size=hidden_size,
                                          num_heads=num_heads,
@@ -36,7 +36,7 @@ class Block1D(nn.Module):
                                          is_decoder=is_decoder,
                                          disable_past_cache=disable_past_cache)
 
-        self.norm2 = LayerNorm1D(hidden_size, eps=layernorm_epsilon)
+        self.norm2 = LayerNorm1D(hidden_size, eps=layernorm_epsilon, dtype=dtype)
 
         self.mlp = MLP1D(hidden_size=hidden_size,
                          mlp_ratio=mlp_ratio,
@@ -45,7 +45,7 @@ class Block1D(nn.Module):
                          bias=bias,
                          disable_past_cache=disable_past_cache)
 
-    def forward(self, hidden_states, attention_mask=None, first_cache = False, seq_lens=None):
+    def forward(self, hidden_states, attention_mask=None, first_cache=False, seq_lens=None):
 
         if not self.apply_post_layernorm:
             residual = hidden_states
@@ -53,9 +53,9 @@ class Block1D(nn.Module):
 
         if self.apply_post_layernorm:
             residual = hidden_states
-        hidden_states = residual + self.attn(hidden_states = hidden_states,
-                                            attention_mask = attention_mask, 
-                                            first_cache=first_cache)
+        hidden_states = residual + self.attn(hidden_states=hidden_states,
+                                             attention_mask=attention_mask,
+                                             first_cache=first_cache)
 
         if not self.apply_post_layernorm:
             residual = hidden_states
@@ -64,7 +64,7 @@ class Block1D(nn.Module):
 
         if self.apply_post_layernorm:
             residual = hidden_states
-        hidden_states = residual + self.mlp(hidden_states = hidden_states,
-                                            first_cache = first_cache)
+        hidden_states = residual + self.mlp(hidden_states=hidden_states,
+                                            first_cache=first_cache)
 
         return hidden_states
